ISS4094: Moved signOut to Settings component and added navigation to home
