+++ charcoal-client/src/App.tsx
@@ -12,16 +12,8 @@ import './App.css';
 import { getConfiguration, getConfigurationError, loadConfiguration, receiveRefreshToken } from './slices/configuration'
 import Spinner from './components/Spinner'
 import { SignInOrUp } from './components/SignIn'
-import { setIntent as setLifeLineIntent } from './slices/lifeLine'
-import { setIntent as setEphemeraIntent } from './slices/ephemera'
-import { setIntent as setPlayerIntent } from './slices/player'
-import { clear as clearPersonalAssets } from './slices/personalAssets'
-import { clear as clearMessages } from './slices/messages'
-import { clear as clearPerceptionCache } from './slices/perceptionCache'
-import { clear as clearActiveCharacters } from './slices/activeCharacters'
-import { clear as clearNavigationTabs } from './slices/UI/navigationTabs'
-import { heartbeat } from './slices/stateSeekingMachine/ssmHeartbeat'
 import useStateSeekingMachines from './components/useSSM';
+import { useNavigate } from 'react-router-dom';
 
 declare module '@mui/styles' {
   interface DefaultTheme extends Theme {}
@@ -29,24 +21,11 @@ declare module '@mui/styles' {
 
 const theme = createTheme();
 
-const signOut = (dispatch) => {
-  dispatch(receiveRefreshToken(undefined))
-  dispatch(setLifeLineIntent(['SIGNOUT']))
-  dispatch(setEphemeraIntent(['SIGNOUT']))
-  dispatch(setPlayerIntent(['SIGNOUT']))
-  dispatch(clearPersonalAssets())
-  dispatch(clearMessages())
-  dispatch(clearPerceptionCache())
-  dispatch(clearActiveCharacters())
-  dispatch(clearNavigationTabs())
-  dispatch(heartbeat)
-}
-
-export const App = ({ signOut }: { signOut: () => void }) => (
+export const App = () => (
   <StyledEngineProvider injectFirst>
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <AppController signOut={signOut} />
+      <AppController />
     </ThemeProvider>
   </StyledEngineProvider>
 )
@@ -71,7 +50,7 @@ const ConfiguredApp = () => {
       return <SignInOrUp />
     }
     else {
-      return <App signOut={() => { dispatch(signOut) }}/>
+      return <App />
     }
   }
   else {
